hw/rdc_setup.sh: add bf_switchd.c
The code for locating the Credo/Avago firmware is disabled in
bf_switchd.c, which leads to a non-working system when the procedure
to merge with the RDC is applied. Currently, this is the only
difference between the open-p4studio and RDC versions of the file,
hence it is safe to simply add bf_switchd.c to the list of files to
copy.

Signed-off-by: Alexander Gall <alexander.gall@gmail.com>